pulley_interpreter

Module decode

Decoding support for pulley bytecode.

Modules

operands

Functions for decoding the operands of an instruction, assuming the opcode has already been decoded.

Structs

Decoder

A Pulley bytecode decoder.

SafeBytecodeStream

A 100% safe implementation of a bytecode stream.

SequencedVisitor

An OpVisitor combinator to sequence one visitor and then another.

UnsafeBytecodeStream

An unsafe bytecode stream.

Enums

DecodingError

An error when decoding Pulley bytecode.

Uninhabited

An uninhabited type that cannot be constructed at runtime.

Traits

BytecodeStream

An abstraction over any kind of bytecode stream.

Decode

Anything that can be decoded from a bytecode stream, e.g. opcodes, immediates, registers, etc…

ExtendedOpVisitor

Like OpVisitor but for extended operations.

OpVisitor

Callbacks upon decoding instructions from bytecode.

Type Aliases

Result

Either an Ok(T) or an Err(DecodingError).