pulley_interpreter::decode

Module operands

Functions for decoding the operands of an instruction, assuming the opcode has already been decoded.

Functions

bitcast_float_from_int_32

bitcast_float_from_int_64

bitcast_int_from_float_32

bitcast_int_from_float_64

br_if32

br_if_not32

br_if_xeq32

br_if_xeq64

br_if_xeq32_i8

br_if_xeq32_i32

br_if_xeq64_i8

br_if_xeq64_i32

br_if_xneq32

br_if_xneq64

br_if_xneq32_i8

br_if_xneq32_i32

br_if_xneq64_i8

br_if_xneq64_i32

br_if_xsgt32_i8

br_if_xsgt32_i32

br_if_xsgt64_i8

br_if_xsgt64_i32

br_if_xsgteq32_i8

br_if_xsgteq32_i32

br_if_xsgteq64_i8

br_if_xsgteq64_i32

br_if_xslt32

br_if_xslt64

br_if_xslt32_i8

br_if_xslt32_i32

br_if_xslt64_i8

br_if_xslt64_i32

br_if_xslteq32

br_if_xslteq64

br_if_xslteq32_i8

br_if_xslteq32_i32

br_if_xslteq64_i8

br_if_xslteq64_i32

br_if_xugt32_u8

br_if_xugt32_u32

br_if_xugt64_u8

br_if_xugt64_u32

br_if_xugteq32_u8

br_if_xugteq32_u32

br_if_xugteq64_u8

br_if_xugteq64_u32

br_if_xult32

br_if_xult64

br_if_xult32_u8

br_if_xult32_u32

br_if_xult64_u8

br_if_xult64_u32

br_if_xulteq32

br_if_xulteq64

br_if_xulteq32_u8

br_if_xulteq32_u32

br_if_xulteq64_u8

br_if_xulteq64_u32

br_table32

bswap32

bswap64

call

call1

call2

call3

call4

call_indirect

call_indirect_host

extended

Decode an extended opcode from pc to match the payload of the “extended” opcode.

f32_from_f64

f32_from_x32_s

f32_from_x32_u

f32_from_x64_s

f32_from_x64_u

f64_from_f32

f64_from_x32_s

f64_from_x32_u

f64_from_x64_s

f64_from_x64_u

fabs32

fabs64

fadd32

fadd64

fceil32

fceil64

fconst32

fconst64

fcopysign32

fcopysign64

fdiv32

fdiv64

feq32

feq64

fextractv32x4

fextractv64x2

ffloor32

ffloor64

fload32be_o32

fload32le_g32

fload32le_o32

fload32le_z

fload64be_o32

fload64le_g32

fload64le_o32

fload64le_z

flt32

flt64

flteq32

flteq64

fmaximum32

fmaximum64

fminimum32

fminimum64

fmov

fmul32

fmul64

fnearest32

fnearest64

fneg32

fneg64

fneq32

fneq64

fselect32

fselect64

fsqrt32

fsqrt64

fstore32be_o32

fstore32le_g32

fstore32le_o32

fstore32le_z

fstore64be_o32

fstore64le_g32

fstore64le_o32

fstore64le_z

fsub32

fsub64

ftrunc32

ftrunc64

jump

nop

pop_frame

pop_frame_restore

push_frame

push_frame_save

ret

sext8

sext16

sext32

stack_alloc32

stack_free32

trap

vabs8x16

vabs16x8

vabs32x4

vabs64x2

vabsf32x4

vabsf64x2

vaddf32x4

vaddf64x2

vaddi8x16

vaddi8x16_sat

vaddi16x8

vaddi16x8_sat

vaddi32x4

vaddi64x2

vaddpairwisei16x8_s

vaddpairwisei32x4_s

vaddu8x16_sat

vaddu16x8_sat

valltrue8x16

valltrue16x8

valltrue32x4

valltrue64x2

vanytrue8x16

vanytrue16x8

vanytrue32x4

vanytrue64x2

vavground8x16

vavground16x8

vband128

vbitmask8x16

vbitmask16x8

vbitmask32x4

vbitmask64x2

vbitselect128

vbnot128

vbor128

vbxor128

vceil32x4

vceil64x2

vconst128

vdivf32x4

vdivf64x2

veq8x16

veq16x8

veq32x4

veq64x2

veqf32x4

veqf64x2

vf32x4_from_i32x4_s

vf32x4_from_i32x4_u

vf64x2_from_i64x2_s

vf64x2_from_i64x2_u

vfdemote

vfloor32x4

vfloor64x2

vfma32x4

vfma64x2

vfpromotelow

vi32x4_from_f32x4_s

vi32x4_from_f32x4_u

vi64x2_from_f64x2_s

vi64x2_from_f64x2_u

vinsertf32

vinsertf64

vinsertx8

vinsertx16

vinsertx32

vinsertx64

vload8x8_s_z

vload8x8_u_z

vload16x4le_s_z

vload16x4le_u_z

vload32x2le_s_z

vload32x2le_u_z

vload128le_g32

vload128le_o32

vload128le_z

vlteqf32x4

vlteqf64x2

vltf32x4

vltf64x2

vmax8x16_s

vmax8x16_u

vmax16x8_s

vmax16x8_u

vmax32x4_s

vmax32x4_u

vmaximumf32x4

vmaximumf64x2

vmin8x16_s

vmin8x16_u

vmin16x8_s

vmin16x8_u

vmin32x4_s

vmin32x4_u

vminimumf32x4

vminimumf64x2

vmov

vmulf32x4

vmulf64x2

vmuli8x16

vmuli16x8

vmuli32x4

vmuli64x2

vnarrow16x8_s

vnarrow16x8_u

vnarrow32x4_s

vnarrow32x4_u

vnarrow64x2_s

vnarrow64x2_u

vnearest32x4

vnearest64x2

vneg8x16

vneg16x8

vneg32x4

vneg64x2

vnegf32x4

vnegf64x2

vneq8x16

vneq16x8

vneq32x4

vneq64x2

vneqf32x4

vneqf64x2

vpopcnt8x16

vqmulrsi16x8

vselect

vshli8x16

vshli16x8

vshli32x4

vshli64x2

vshri8x16_s

vshri8x16_u

vshri16x8_s

vshri16x8_u

vshri32x4_s

vshri32x4_u

vshri64x2_s

vshri64x2_u

vshuffle

vslt8x16

vslt16x8

vslt32x4

vslt64x2

vslteq8x16

vslteq16x8

vslteq32x4

vslteq64x2

vsplatf32

vsplatf64

vsplatx8

vsplatx16

vsplatx32

vsplatx64

vsqrt32x4

vsqrt64x2

vstore128le_g32

vstore128le_o32

vstore128le_z

vsubf32x4

vsubf64x2

vsubi8x16

vsubi8x16_sat

vsubi16x8

vsubi16x8_sat

vsubi32x4

vsubi64x2

vsubu8x16_sat

vsubu16x8_sat

vswizzlei8x16

vtrunc32x4

vtrunc64x2

vult8x16

vult16x8

vult32x4

vult64x2

vulteq8x16

vulteq16x8

vulteq32x4

vulteq64x2

vunarrow64x2_u

vwidenhigh8x16_s

vwidenhigh8x16_u

vwidenhigh16x8_s

vwidenhigh16x8_u

vwidenhigh32x4_s

vwidenhigh32x4_u

vwidenlow8x16_s

vwidenlow8x16_u

vwidenlow16x8_s

vwidenlow16x8_u

vwidenlow32x4_s

vwidenlow32x4_u

x32_from_f32_s

x32_from_f32_s_sat

x32_from_f32_u

x32_from_f32_u_sat

x32_from_f64_s

x32_from_f64_s_sat

x32_from_f64_u

x32_from_f64_u_sat

x64_from_f32_s

x64_from_f32_s_sat

x64_from_f32_u

x64_from_f32_u_sat

x64_from_f64_s

x64_from_f64_s_sat

x64_from_f64_u

x64_from_f64_u_sat

xabs32

xabs64

xadd32

xadd64

xadd32_u8

xadd32_u32

xadd32_uoverflow_trap

xadd64_u8

xadd64_u32

xadd64_uoverflow_trap

xadd128

xband32

xband64

xband32_s8

xband32_s32

xband64_s8

xband64_s32

xbmask32

xbmask64

xbnot32

xbnot64

xbor32

xbor64

xbor32_s8

xbor32_s32

xbor64_s8

xbor64_s32

xbxor32

xbxor64

xbxor32_s8

xbxor32_s32

xbxor64_s8

xbxor64_s32

xclz32

xclz64

xconst8

xconst16

xconst32

xconst64

xctz32

xctz64

xdiv32_s

xdiv32_u

xdiv64_s

xdiv64_u

xeq32

xeq64

xextractv8x16

xextractv16x8

xextractv32x4

xextractv64x2

xjump

xload8_s32_g32

xload8_s32_g32bne

xload8_s32_o32

xload8_s32_z

xload8_u32_g32

xload8_u32_g32bne

xload8_u32_o32

xload8_u32_z

xload16be_s32_o32

xload16be_u32_o32

xload16le_s32_g32

xload16le_s32_g32bne

xload16le_s32_o32

xload16le_s32_z

xload16le_u32_g32

xload16le_u32_g32bne

xload16le_u32_o32

xload16le_u32_z

xload32be_o32

xload32le_g32

xload32le_g32bne

xload32le_o32

xload32le_z

xload64be_o32

xload64le_g32

xload64le_g32bne

xload64le_o32

xload64le_z

xmadd32

xmadd64

xmax32_s

xmax32_u

xmax64_s

xmax64_u

xmin32_s

xmin32_u

xmin64_s

xmin64_u

xmov

xmov_fp

xmov_lr

xmul32

xmul64

xmul32_s8

xmul32_s32

xmul64_s8

xmul64_s32

xmulhi64_s

xmulhi64_u

xneg32

xneg64

xneq32

xneq64

xone

xpopcnt32

xpopcnt64

xrem32_s

xrem32_u

xrem64_s

xrem64_u

xrotl32

xrotl64

xrotr32

xrotr64

xselect32

xselect64

xshl32

xshl64

xshl32_u6

xshl64_u6

xshr32_s

xshr32_s_u6

xshr32_u

xshr32_u_u6

xshr64_s

xshr64_s_u6

xshr64_u

xshr64_u_u6

xslt32

xslt64

xslteq32

xslteq64

xstore8_g32

xstore8_g32bne

xstore8_o32

xstore8_z

xstore16be_o32

xstore16le_g32

xstore16le_g32bne

xstore16le_o32

xstore16le_z

xstore32be_o32

xstore32le_g32

xstore32le_g32bne

xstore32le_o32

xstore32le_z

xstore64be_o32

xstore64le_g32

xstore64le_g32bne

xstore64le_o32

xstore64le_z

xsub32

xsub64

xsub32_u8

xsub32_u32

xsub64_u8

xsub64_u32

xsub128

xult32

xult64

xulteq32

xulteq64

xwidemul64_s

xwidemul64_u

xzero

zext8

zext16

zext32