pulley_interpreter::disas

Struct Disassembler

pub struct Disassembler<'a> { /* private fields */ }

A Pulley bytecode disassembler.

This is implemented as an OpVisitor, where you pass a Disassembler to a Decoder in order to disassemble instructions from a bytecode stream.

Alternatively, you can use the Disassembler::disassemble_all method to disassemble a complete bytecode stream.

Implementations

impl<'a> Disassembler<'a>

pub fn disassemble_all(bytecode: &'a [u8]) -> Result<String>

Disassemble every instruction in the given bytecode stream.

pub fn new(bytecode: &'a [u8]) -> Self

Create a new Disassembler that can be used to incrementally disassemble instructions from the given bytecode stream.

pub fn offsets(&mut self, offsets: bool) -> &mut Self

Whether to prefix each instruction’s disassembly with its offset.

True by default.

pub fn hexdump(&mut self, hexdump: bool) -> &mut Self

Whether to include a hexdump of the bytecode in the disassembly.

True by default.

pub fn br_tables(&mut self, enable: bool) -> &mut Self

Whether to include branche tables in the disassembly.

True by default.

pub fn start_offset(&mut self, offset: usize) -> &mut Self

Configures the offset that this function starts from, if it doesn’t start from 0.

This can possibly be useful when a single function at a time is being disassembled.

pub fn disas(&self) -> &str

Get the disassembly thus far.