pulley_interpreter

Module interp

Interpretation of pulley bytecode.

Structs

FRegVal

An f register value: floats.

MachineState

The machine state for a Pulley virtual machine: the various registers and stack.

VRegVal

A v register value: vectors.

Vm

A virtual machine for interpreting Pulley bytecode.

XRegVal

An x register value: integers.

Enums

DoneReason

Reason that the pulley interpreter has ceased execution.

RegType

The type of a register in the Pulley machine state.

TrapKind

Stored within DoneReason::Trap.

Val

A value that can be stored in a register.