pulley_interpreter::interp

Struct Vm

pub struct Vm { /* private fields */ }

A virtual machine for interpreting Pulley bytecode.

Implementations

impl Vm

pub fn new() -> Self

Create a new virtual machine with the default stack size.

pub fn with_stack(stack_size: usize) -> Self

Create a new virtual machine with the given stack.

pub fn state(&self) -> &MachineState

Get a shared reference to this VM’s machine state.

pub fn state_mut(&mut self) -> &mut MachineState

Get an exclusive reference to this VM’s machine state.

pub unsafe fn call<'a, T>( &'a mut self, func: NonNull<u8>, args: &[Val], rets: T, ) -> DoneReason<impl Iterator<Item = Val> + use<'a, T>>
where T: IntoIterator<Item = RegType> + 'a,

Call a bytecode function.

The given func must point to the beginning of a valid Pulley bytecode function.

The given args must match the number and type of arguments that function expects.

The given rets must match the function’s actual return types.

Returns either the resulting values, or the PC at which a trap was raised.

pub unsafe fn call_start<'a>(&'a mut self, args: &[Val]) -> *mut u8

Peforms the initial part of Vm::call in setting up the args provided in registers according to Pulley’s ABI.

Return

Returns the old lr register value. The current lr value is replaced with a sentinel that triggers a return to the host when returned-to.

Unsafety

All the same unsafety as call and additiionally, you must invoke call_run and then call_end after calling call_start. If you don’t want to wrangle these invocations, use call instead of call_{start,run,end}.

pub unsafe fn call_run(&mut self, pc: NonNull<u8>) -> DoneReason<()>

Peforms the internal part of Vm::call where bytecode is actually executed.

Unsafety

In addition to all the invariants documented for call, you may only invoke call_run after invoking call_start to initialize this call’s arguments.

pub unsafe fn call_end<'a>( &'a mut self, old_ret: *mut u8, rets: impl IntoIterator<Item = RegType> + 'a, ) -> impl Iterator<Item = Val> + 'a

Peforms the tail end of Vm::call by returning the values as determined by rets according to Pulley’s ABI.

The old_ret value should have been provided from call_start previously.

Unsafety

In addition to the invariants documented for call, this may only be called after call_run.

pub fn fp(&self) -> *mut u8

Returns the current fp register value.

pub fn lr(&self) -> *mut u8

Returns the current lr register value.

pub unsafe fn set_fp(&mut self, fp: *mut u8)

Sets the current fp register value.

pub unsafe fn set_lr(&mut self, lr: *mut u8)

Sets the current lr register value.