pulley_interpreter::regs

Trait Reg

pub trait Reg:
    Sized
    + Copy
    + Eq
    + Ord
    + Hash
    + Into<AnyReg>
    + Debug
    + Display {
    const RANGE: Range<u8>;

    // Required methods
    unsafe fn new_unchecked(index: u8) -> Self;
    fn to_u8(self) -> u8;

    // Provided methods
    fn new(index: u8) -> Option<Self> { ... }
    fn index(self) -> usize { ... }
}

Trait for common register operations.

Required Associated Constants

const RANGE: Range<u8>

Range of valid register indices.

Required Methods

unsafe fn new_unchecked(index: u8) -> Self

Convert a register index to a register, without bounds checking.

fn to_u8(self) -> u8

Convert a register to its index.

Provided Methods

fn new(index: u8) -> Option<Self>

Convert a register index to a register, with bounds checking.

fn index(self) -> usize

Convert a register to its index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.